+++ test/lib/Streamly/Test/Prelude/Common.hs
@@ -30,7 +30,7 @@ module Streamly.Test.Prelude.Common
     , constructWithCons
     , constructWithConsM
     , constructWithYield
-    , constructWithYieldM
+    , constructWithFromEffect
     , simpleOps
     -- * Applicative operations
     , applicativeOps
@@ -415,7 +415,7 @@ constructWithYield listT op len =
         let list = replicate (fromIntegral len) 0
         listEquals (==) (listT strm) list
 
-constructWithYieldM ::
+constructWithFromEffect ::
        (IsStream t
 #if __GLASGOW_HASKELL__ < 806
        , Monoid (t IO Int)
@@ -425,13 +425,13 @@ constructWithYieldM ::
     -> (t IO Int -> SerialT IO Int)
     -> Word8
     -> Property
-constructWithYieldM listT op len =
+constructWithFromEffect listT op len =
     withMaxSuccess maxTestCount $
     monadicIO $ do
         strm <-
             run
                 $ S.toList . op . S.take (fromIntegral len)
-                $ foldMap S.yieldM (repeat (return 0))
+                $ foldMap S.fromEffect (repeat (return 0))
         let list = replicate (fromIntegral len) 0
         listEquals (==) (listT strm) list
 
@@ -447,7 +447,7 @@ simpleProps constr op a = monadicIO $ do
 simpleOps :: IsStream t => (t IO Int -> SerialT IO Int) -> Spec
 simpleOps op = do
   prop "fromPure a = a" $ simpleProps S.fromPure op
-  prop "yieldM a = a" $ simpleProps (S.yieldM . return) op
+  prop "fromEffect a = a" $ simpleProps (S.fromEffect . return) op
 
 -------------------------------------------------------------------------------
 -- Applicative operations
@@ -786,12 +786,12 @@ loops t tsrt hsrt = do
     where
         loopHead x = do
             -- this print line is important for the test (causes a bind)
-            S.yieldM $ putStrLn "LoopHead..."
+            S.fromEffect $ putStrLn "LoopHead..."
             t $ (if x < 3 then loopHead (x + 1) else nil) <> return x
 
         loopTail x = do
             -- this print line is important for the test (causes a bind)
-            S.yieldM $ putStrLn "LoopTail..."
+            S.fromEffect $ putStrLn "LoopTail..."
             t $ return x <> (if x < 3 then loopTail (x + 1) else nil)
 
 ---------------------------------------------------------------------------
@@ -1385,7 +1385,7 @@ parallelCheck t f = do
         (S.toList . t) (event 4 `f` (event 3 `f` (event 2 `f` event 1)))
             `shouldReturn` [1..4]
 
-    where event n = S.yieldM (threadDelay (n * 200000)) >> return n
+    where event n = S.fromEffect (threadDelay (n * 200000)) >> return n
 
 -------------------------------------------------------------------------------
 -- Exception ops